fix: Take deployed CRDs into account when fixing namespaces
When a deployment also contains CRDs, the k8s client will not know if
an object is namespaced before the CRD is actually applied. This leads
to flaky behaviour between the initial and all other deployments.

This commit delays fixing of namespaces until all objects are rendered.
Then it takes CRDs into account when determining if an object is namespaced
or not.