fix: Reimplement GetFilteredPreferredGVKs to account for resources without PV
Sometimes individual resources inside a group do not have an APIResource
with a version that matches the preferred version of the group. In that
case, we must use the latest version instead.

Before this, such resources were completely ignored/skipped, leading to
orphan detection to not work for these resources.